restructured arithmetic, added overload methods and tests
