+++ tests/test_signal_arithmetic.py
@@ -2,8 +2,8 @@ import numpy as np
 import numpy.testing as npt
 from pytest import raises
 
+import pyfar.signal as signal
 from pyfar.signal import Signal
-import pyfar.arithmetic as arithmetic
 
 
 # test adding two Signals
@@ -12,7 +12,7 @@ def test_add_two_signals_time():
     x = Signal([1, 0, 0], 44100)
 
     # time domain
-    y = arithmetic.add((x, x), 'time')
+    y = signal.add((x, x), 'time')
     # check if old signal did not change
     npt.assert_allclose(x.time, np.atleast_2d([1, 0, 0]), atol=1e-15)
     # check result
@@ -27,7 +27,7 @@ def test_add_two_signals_freq():
     x = Signal([1, 0, 0], 44100)
 
     # frequency domain
-    y = arithmetic.add((x, x), 'freq')
+    y = signal.add((x, x), 'freq')
     # check if old signal did not change
     npt.assert_allclose(x.time, np.atleast_2d([1, 0, 0]), atol=1e-15)
     # check result
@@ -36,10 +36,11 @@ def test_add_two_signals_freq():
     npt.assert_allclose(y.freq, np.atleast_2d([2, 2]), atol=1e-15)
 
 
+# test adding three signals
 def test_add_three_signals():
     # generate and add signals
     x = Signal([1, 0, 0], 44100)
-    y = arithmetic.add((x, x, x), 'time')
+    y = signal.add((x, x, x), 'time')
 
     # check if old signal did not change
     npt.assert_allclose(x.time, np.atleast_2d([1, 0, 0]), atol=1e-15)
@@ -54,7 +55,7 @@ def test_add_three_signals():
 def test_add_signal_and_number():
     # generate and add signals
     x = Signal([1, 0, 0], 44100)
-    y = arithmetic.add((x, 1), 'time')
+    y = signal.add((x, 1), 'time')
 
     # check if old signal did not change
     npt.assert_allclose(x.time, np.atleast_2d([1, 0, 0]), atol=1e-15)
@@ -69,7 +70,7 @@ def test_add_signal_and_number():
 def test_add_number_and_signal():
     # generate and add signals
     x = Signal([1, 0, 0], 44100)
-    y = arithmetic.add((1, x), 'time')
+    y = signal.add((1, x), 'time')
 
     # check if old signal did not change
     npt.assert_allclose(x.time, np.atleast_2d([1, 0, 0]), atol=1e-15)
@@ -80,6 +81,67 @@ def test_add_number_and_signal():
     npt.assert_allclose(y.time, np.atleast_2d([2, 1, 1]), atol=1e-15)
 
 
+def test_subtraction():
+    # only test one case - everything else is tested below
+    x = Signal([1, 0, 0], 44100)
+    y = Signal([0, 1, 0], 44100)
+    z = signal.subtract((x, y), 'time')
+
+    # check result
+    npt.assert_allclose(z.time, np.atleast_2d([1, -1, 0]), atol=1e-15)
+
+
+def test_multiplication():
+    # only test one case - everything else is tested below
+    x = Signal([1, 0, 0], 44100)
+    y = Signal([0, 1, 0], 44100)
+    z = signal.multiply((x, y), 'time')
+
+    # check result
+    npt.assert_allclose(z.time, np.atleast_2d([0, 0, 0]), atol=1e-15)
+
+
+def test_division():
+    # only test one case - everything else is tested below
+    x = Signal([1, 0, 0], 44100)
+    y = Signal([2, 2, 2], 44100)
+    z = signal.divide((x, y), 'time')
+
+    # check result
+    npt.assert_allclose(z.time, np.atleast_2d([0.5, 0, 0]), atol=1e-15)
+
+
+def test_power():
+    # only test one case - everything else is tested below
+    x = Signal([2, 1, 0], 44100)
+    y = Signal([2, 2, 2], 44100)
+    z = signal.power((x, y), 'time')
+
+    # check result
+    npt.assert_allclose(z.time, np.atleast_2d([4, 1, 0]), atol=1e-15)
+
+
+def test_overloaded_operators():
+    x = Signal([2, 1, 0], 44100, n_samples=5, domain='freq')
+    y = Signal([2, 2, 2], 44100, n_samples=5, domain='freq')
+
+    # addition
+    z = x + y
+    npt.assert_allclose(z.freq, np.array([4, 3, 2], ndmin=2), atol=1e-15)
+    # subtraction
+    z = x - y
+    npt.assert_allclose(z.freq, np.array([0, -1, -2], ndmin=2), atol=1e-15)
+    # multiplication
+    z = x * y
+    npt.assert_allclose(z.freq, np.array([4, 2, 0], ndmin=2), atol=1e-15)
+    # division
+    z = x / y
+    npt.assert_allclose(z.freq, np.array([1, .5, 0], ndmin=2), atol=1e-15)
+    # power
+    z = x**y
+    npt.assert_allclose(z.freq, np.array([4, 1, 0], ndmin=2), atol=1e-15)
+
+
 def test_assert_match_for_arithmetic():
     s = Signal([1, 2, 3, 4], 44100)
     s1 = Signal([1, 2, 3, 4], 48000)
@@ -88,48 +150,48 @@ def test_assert_match_for_arithmetic():
     s5 = Signal([1, 2, 3, 4], 44100, signal_type="power", fft_norm='amplitude')
 
     # check with two signals
-    arithmetic._assert_match_for_arithmetic((s, s), 'time')
+    signal._assert_match_for_arithmetic((s, s), 'time')
     # check with one signal and one array like
-    arithmetic._assert_match_for_arithmetic((s, [1, 2]), 'time')
+    signal._assert_match_for_arithmetic((s, [1, 2]), 'time')
     # check with more than two inputs
-    arithmetic._assert_match_for_arithmetic((s, s, s), 'time')
+    signal._assert_match_for_arithmetic((s, s, s), 'time')
 
     # check output
-    out = arithmetic._assert_match_for_arithmetic((s, s), 'time')
+    out = signal._assert_match_for_arithmetic((s, s), 'time')
     assert out[0] == 44100
     assert out[1] == 4
     assert out[2] == 'energy'
-    out = arithmetic._assert_match_for_arithmetic((s, s4), 'time')
+    out = signal._assert_match_for_arithmetic((s, s4), 'time')
     assert out[2] == 'power'
-    out = arithmetic._assert_match_for_arithmetic((s5, s4), 'time')
+    out = signal._assert_match_for_arithmetic((s5, s4), 'time')
     assert out[2] == 'power'
     assert out[3] == 'amplitude'
 
     # check with only one argument
     with raises(TypeError):
-        arithmetic._assert_match_for_arithmetic((s, s))
+        signal._assert_match_for_arithmetic((s, s))
     # check with single input
     with raises(ValueError):
-        arithmetic._assert_match_for_arithmetic(s, 'time')
+        signal._assert_match_for_arithmetic(s, 'time')
     # check with invalid data type
     with raises(ValueError):
-        arithmetic._assert_match_for_arithmetic((s, ['str', 'ing']), 'time')
+        signal._assert_match_for_arithmetic((s, ['str', 'ing']), 'time')
     # check with complex data and time domain signal
     with raises(ValueError):
-        arithmetic._assert_match_for_arithmetic(
+        signal._assert_match_for_arithmetic(
             (s, np.array([1 + 1j])), 'time')
     # test signals with different sampling rates
     with raises(ValueError):
-        arithmetic._assert_match_for_arithmetic((s, s1), 'time')
+        signal._assert_match_for_arithmetic((s, s1), 'time')
     # test signals with different n_samples
     with raises(ValueError):
-        arithmetic._assert_match_for_arithmetic((s, s2), 'time')
+        signal._assert_match_for_arithmetic((s, s2), 'time')
 
 
 def test_get_arithmetic_data_with_array():
 
     data_in = np.asarray(1)
-    data_out = arithmetic._get_arithmetic_data(data_in, None, None)
+    data_out = signal._get_arithmetic_data(data_in, None, None)
     npt.assert_allclose(data_in, data_out)
 
 
@@ -160,7 +222,7 @@ def test_get_arithmetic_data_with_signal():
             print(f"Testing from {', '.join(m_in)} to {domain}.")
 
             # get output data
-            data_out = arithmetic._get_arithmetic_data(
+            data_out = signal._get_arithmetic_data(
                 s_in, n_samples=3, domain=domain)
             if domain == 'time':
                 npt.assert_allclose(s_ref.time, data_out, atol=1e-15)
@@ -170,5 +232,5 @@ def test_get_arithmetic_data_with_signal():
 
 def test_get_arithmetic_data_assertion():
     with raises(ValueError):
-        arithmetic._get_arithmetic_data(
+        signal._get_arithmetic_data(
             Signal(1, 44100), 1, 'space')
